+++ NEWS.md
@@ -1,4 +1,4 @@
-## packer 0.1.2.9000
+## packer 0.1.2
 
 - Few typo in documentation. 
 - New `framework7` parameter for `golem_scaffold` to create plug and play Framework7 powered Shiny app.